chore: split last seen column into "Status" and "Last Heartbeat" columns.
