+++ scripts/generate_sites.sh
@@ -41,8 +41,8 @@ for i in $(seq 1 $ENTRIES); do
     MEMORY_USAGE=$(( (SEED * 12345) % (MEMORY_TOTAL_AMOUNT + 1) ))  # 0 to MEMORY_TOTAL_AMOUNT
     DISK_TOTAL_SIZE=2000000000000 #2TB
     DISK_USAGE=$(( (SEED * 12345678) % (DISK_TOTAL_SIZE + 1) ))  # 0 to DISK_TOTAL_SIZE
-    # Generates a random date and time between date.now and 10 minutes ago.
-    LAST_UPDATED_AT=$(date -u -d "@$(( $(date -u -d '10 minutes ago' +%s) + SEED % ($(date -u +%s) - $(date -u -d '10 minutes ago' +%s)) ))" +'%Y-%m-%d %H:%M:%S')
+    # Generates a random date and time between date.now and 30 minutes ago.
+    LAST_UPDATED_AT=$(date -u -d "@$(( $(date -u -d '30 minutes ago' +%s) + SEED % ($(date -u +%s) - $(date -u -d '30 minutes ago' +%s)) ))" +'%Y-%m-%d %H:%M:%S')
     INSTANCE_STATUSES=$(printf "$INSTANCE_STATUSES_TEMPLATE" $((SEED % 200 + 1)))
     MEMBER_STATUSES=$(printf "$MEMBER_STATUSES_TEMPLATE" $((SEED % 4)))
     SITE_DETAILS_VALUES+=("($CORE_SITE_ID, '$STATUS', '$CPU_TOTAL_COUNT', '$CPU_LOAD_1', '$CPU_LOAD_5', '$CPU_LOAD_15', '$MEMORY_TOTAL_AMOUNT', '$MEMORY_USAGE', '$DISK_TOTAL_SIZE', '$DISK_USAGE', '$INSTANCE_COUNT', '$INSTANCE_STATUSES', '$MEMBER_COUNT', '$MEMBER_STATUSES', '$LAST_UPDATED_AT')")
