+++ ui/src/pages/clusters/metrics/ClusterCpu.tsx
